+++ gui/custom_widgets/report_editor.h
@@ -9,7 +9,7 @@
 #include "highlighter.h"
 #include "completer.h"
 #include "text_editor.h"
-#include "util/print.h"
+#include "print/print.h"
 #include "main_window_abstract.h"
